refactor styles modules
