+++ apps/nextjs-website/src/components/molecules/CodeBlockPart/CodeBlockPart.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Box, Typography, useTheme } from '@mui/material';
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
 import { lightCustomStyle } from '@/components/molecules/CodeBlockPart/lightCustomStyle';
-import styles from './CodeBlockPart.module.css';
+import styles from './styles.module.css';
 import { darkCustomStyle } from '@/components/molecules/CodeBlockPart/darkCustomStyle';
 import { translations } from '@/_contents/translations';
 import CopyToClipboard from '@/components/atoms/CopyToClipboard/CopyToClipboard';
